feat: variant page section update
