+++ packages/sections/src/variant/VariantEffectPredictor/Body.tsx
@@ -1,10 +1,10 @@
 import { useQuery } from "@apollo/client";
 import { Box } from "@mui/material";
-import { Link, SectionItem, DataTable, Tooltip } from "ui";
+import { Link, SectionItem, Tooltip, OtTable } from "ui";
 import { Fragment } from "react";
 import { definition } from "../VariantEffectPredictor";
 import Description from "../VariantEffectPredictor/Description";
-import { defaultRowsPerPageOptions, naLabel } from "../../constants";
+import { naLabel } from "../../constants";
 import { identifiersOrgLink } from "../../utils/global";
 import VARIANT_EFFECT_PREDICTOR_QUERY from "./VariantEffectPredictorQuery.gql";
 
@@ -19,17 +19,16 @@ const columns = [
     comparator: (a, b) => a.transcriptIndex - b.transcriptIndex,
     renderCell: ({ target, transcriptId }) => {
       if (!target) return naLabel;
-      let displayElement = <Link to={`../target/${target.id}`}>{target.approvedSymbol}</Link>
+      let displayElement = <Link to={`../target/${target.id}`}>{target.approvedSymbol}</Link>;
       if (transcriptId) {
-        displayElement = 
+        displayElement = (
           <Tooltip
             title={
               <Box>
                 Ensembl canonical transcript:{" "}
                 <Link
                   external
-                  to={`https://www.ensembl.org/Homo_sapiens/Transcript/Summary?db=core;g=${
-                      target.id};t=${transcriptId}`}
+                  to={`https://www.ensembl.org/Homo_sapiens/Transcript/Summary?db=core;g=${target.id};t=${transcriptId}`}
                 >
                   {transcriptId}
                 </Link>
@@ -39,14 +38,15 @@ const columns = [
           >
             {displayElement}
           </Tooltip>
+        );
       }
       return displayElement;
-    }
+    },
   },
   {
     id: "variantConsequences.label",
     label: "Predicted consequence",
-    renderCell: ({ variantConsequences }) => (
+    renderCell: ({ variantConsequences }) =>
       variantConsequences.length
         ? variantConsequences.map(({ id, label }, i, arr) => (
             <Fragment key={id}>
@@ -56,12 +56,13 @@ const columns = [
               {i < arr.length - 1 && ", "}
             </Fragment>
           ))
-        : naLabel
-    ),
+        : naLabel,
     exportValue: ({ variantConsequences }) => {
-      return variantConsequences.map(({ label }) => {
-        return formatVariantConsequenceLabel(label)
-      }).join(", ")
+      return variantConsequences
+        .map(({ label }) => {
+          return formatVariantConsequenceLabel(label);
+        })
+        .join(", ");
     },
   },
   {
@@ -90,29 +91,27 @@ const columns = [
   {
     id: "uniprotAccession",
     label: "Uniprot accession",
-    renderCell: ({ uniprotAccessions }) => (
+    renderCell: ({ uniprotAccessions }) =>
       uniprotAccessions?.length
         ? uniprotAccessions.map((id, i, arr) => (
             <Fragment key={id}>
               <Link external to={`https://identifiers.org/uniprot:${id}`}>
                 {id}
               </Link>
-              { i < arr.length - 1 && ", " }
+              {i < arr.length - 1 && ", "}
             </Fragment>
           ))
-        : naLabel
-    ),
-    exportValue: ({ uniprotAccessions }) => (uniprotAccessions ?? []).join(", ")
-  }
+        : naLabel,
+    exportValue: ({ uniprotAccessions }) => (uniprotAccessions ?? []).join(", "),
+  },
 ];
 
 type BodyProps = {
-  id: string,
-  entity: string,
+  id: string;
+  entity: string;
 };
 
 export function Body({ id, entity }: BodyProps) {
-
   const variables = {
     variantId: id,
   };
@@ -126,23 +125,23 @@ export function Body({ id, entity }: BodyProps) {
       definition={definition}
       request={request}
       entity={entity}
-      renderDescription={({ variant }) => (
+      renderDescription={() => (
         <Description
-          variantId={variant.id}
-          referenceAllele={variant.referenceAllele}
-          alternateAllele={variant.alternateAllele}
+          variantId={request.data?.variant.id}
+          referenceAllele={request.data?.variant.referenceAllele}
+          alternateAllele={request.data?.variant.alternateAllele}
         />
       )}
-      renderBody={({ variant }) => {
+      renderBody={() => {
         return (
-          <DataTable
+          <OtTable
             columns={columns}
-            rows={variant.transcriptConsequences}
+            rows={request.data?.variant.transcriptConsequences}
             dataDownloader
-            rowsPerPageOptions={defaultRowsPerPageOptions}
             query={VARIANT_EFFECT_PREDICTOR_QUERY.loc.source.body}
             variables={variables}
             sortBy="target.approvedSymbol"
+            loading={request.loading}
           />
         );
       }}
