+++ packages/ui/src/components/DisplayVariantId.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { ReactNode, useState } from "react";
 import {
   Box,
   Typography,
@@ -21,27 +21,28 @@ type DisplayVariantIdProps = {
   referenceAllele: string;
   alternateAllele: string;
   maxChars?: number;
-  expand?: boolean
+  expand?: boolean;
 };
 
 function DisplayVariantId({
-      variantId: otVariantId,
-      referenceAllele,
-      alternateAllele,
-      maxChars = 6,
-      expand = true,
-    }: DisplayVariantIdProps) {
-
+  variantId: otVariantId,
+  referenceAllele,
+  alternateAllele,
+  maxChars = 6,
+  expand = true,
+}: DisplayVariantIdProps): ReactNode {
   const [open, setOpen] = useState(false);
   const [snackbarOpen, setSnackbarOpen] = useState(false);
 
+  if (!otVariantId || !referenceAllele || !alternateAllele) return null;
+
   function handleClick() {
     setOpen(true);
-  };
+  }
 
   function handleClose() {
     setOpen(false);
-  };
+  }
 
   function handleCloseSnackbar() {
     setSnackbarOpen(false);
@@ -52,18 +53,17 @@ function DisplayVariantId({
     navigator.clipboard.writeText(text);
   }
 
-  const idParts = otVariantId.split('_');
+  const idParts = otVariantId.split("_");
   if (idParts[0] === "OTVAR") {
     idParts.shift();
-  };
+  }
   let isHashed, stem;
-  if (idParts.at(-2) === referenceAllele &&
-      idParts.at(-1) === alternateAllele) {
+  if (idParts.at(-2) === referenceAllele && idParts.at(-1) === alternateAllele) {
     isHashed = false;
-    stem = idParts.slice(0, -2).join('_');
+    stem = idParts.slice(0, -2).join("_");
   } else {
     isHashed = true;
-    stem = idParts.slice(0, -1).join('_');
+    stem = idParts.slice(0, -1).join("_");
   }
 
   const longReferenceAllele = referenceAllele.length > maxChars;
@@ -82,86 +82,82 @@ function DisplayVariantId({
             borderRadius: "0.3em",
             "&:hover": {
               background: expand ? highlightBackground : "transparent",
-            }
+            },
           }}
         >
-          {stem}
+          {stem}_
+          {longReferenceAllele ? (
+            <HighlightBox hlight={expand}>
+              <span style={{ fontSize: "0.94em", fontStyle: "italic" }}>DEL</span>
+            </HighlightBox>
+          ) : (
+            referenceAllele
+          )}
           _
-          {longReferenceAllele
-            ? <HighlightBox hlight={expand}>
-                <span style={{ fontSize: "0.94em", fontStyle: "italic"}}>DEL</span>
-              </HighlightBox>
-            : referenceAllele
-          }
-          _
-          {longAlternateAllele
-            ? <HighlightBox hlight={expand}>
-                <span style={{ fontSize: "0.94em", fontStyle: "italic"}}>INS</span>
-              </HighlightBox>
-            : alternateAllele
-          }
+          {longAlternateAllele ? (
+            <HighlightBox hlight={expand}>
+              <span style={{ fontSize: "0.94em", fontStyle: "italic" }}>INS</span>
+            </HighlightBox>
+          ) : (
+            alternateAllele
+          )}
         </Box>
-        {
-          expand &&
-            <>
-              <Dialog
-                open={open}
-                onClose={handleClose}
-                scroll="paper"
-                aria-labelledby="dialog-title"
-                aria-describedby="dialog-description"
-                maxWidth="md"
-              >
-                <DialogTitle id="dialog-title">
-                  <Typography variant="h6" component="span">
-                    Variant ID
-                  </Typography>
-                  <IconButton
-                    onClick={handleClose}
-                    sx={{
-                      zIndex: "2",
-                      position: "absolute",
-                      top: "0",
-                      right: "0",
-                      padding: "0.7em",
-                    }}
-                  >
-                    <FontAwesomeIcon icon={faXmark} />
-                  </IconButton>
-                </DialogTitle>
-                <DialogContent
-                  dividers={true}
-                  sx={{ padding: "0 1.5em 3em" }}
+        {expand && (
+          <>
+            <Dialog
+              open={open}
+              onClose={handleClose}
+              scroll="paper"
+              aria-labelledby="dialog-title"
+              aria-describedby="dialog-description"
+              maxWidth="md"
+            >
+              <DialogTitle id="dialog-title">
+                <Typography variant="h6" component="span">
+                  Variant ID
+                </Typography>
+                <IconButton
+                  onClick={handleClose}
+                  sx={{
+                    zIndex: "2",
+                    position: "absolute",
+                    top: "0",
+                    right: "0",
+                    padding: "0.7em",
+                  }}
                 >
-                  { isHashed &&
-                      <CopyPanel
-                        label="Hashed Variant ID"
-                        tooltipText="Variant ID used in Open Targets data."
-                        text={otVariantId}
-                        copyToClipboard={copyToClipboard}
-                      />
-                  }
+                  <FontAwesomeIcon icon={faXmark} />
+                </IconButton>
+              </DialogTitle>
+              <DialogContent dividers={true} sx={{ padding: "0 1.5em 3em" }}>
+                {isHashed && (
                   <CopyPanel
-                    label="Full Variant ID"
-                    text={`${stem}_${referenceAllele}_${alternateAllele}`}
+                    label="Hashed Variant ID"
+                    tooltipText="Variant ID used in Open Targets data."
+                    text={otVariantId}
                     copyToClipboard={copyToClipboard}
                   />
-                </DialogContent>
-              </Dialog>
-              <Snackbar
-                open={snackbarOpen}
-                onClose={handleCloseSnackbar}
-                message="Copied to clipboard"
-                autoHideDuration={3000}
-              />
-            </>
-        }
+                )}
+                <CopyPanel
+                  label="Full Variant ID"
+                  text={`${stem}_${referenceAllele}_${alternateAllele}`}
+                  copyToClipboard={copyToClipboard}
+                />
+              </DialogContent>
+            </Dialog>
+            <Snackbar
+              open={snackbarOpen}
+              onClose={handleCloseSnackbar}
+              message="Copied to clipboard"
+              autoHideDuration={3000}
+            />
+          </>
+        )}
       </>
     );
   }
 
   return `${stem}_${referenceAllele}_${alternateAllele}`;
-
 }
 
 function HighlightBox({ children, hlight = true }) {
@@ -188,15 +184,18 @@ type CopyPanelProps = {
 function CopyPanel({ label, text, tooltipText, copyToClipboard }: CopyPanelProps) {
   return (
     <Box mt={2}>
-      {
-        tooltipText
-          ? <Tooltip title={tooltipText} showHelpIcon>
-              <Typography variant="subtitle1" component="span">{label}</Typography>
-            </Tooltip>
-          : <Typography variant="subtitle1">{label}</Typography>
-      }
-      <Box sx={{
-          marginTop: '0.1em',
+      {tooltipText ? (
+        <Tooltip title={tooltipText} showHelpIcon>
+          <Typography variant="subtitle1" component="span">
+            {label}
+          </Typography>
+        </Tooltip>
+      ) : (
+        <Typography variant="subtitle1">{label}</Typography>
+      )}
+      <Box
+        sx={{
+          marginTop: "0.1em",
           backgroundColor: theme => theme.palette.grey[300],
           position: "relative",
         }}
@@ -214,7 +213,7 @@ function CopyPanel({ label, text, tooltipText, copyToClipboard }: CopyPanelProps
         <Typography
           variant="body2"
           sx={{
-            padding: "1em 3.2em 1em 1em",   
+            padding: "1em 3.2em 1em 1em",
             textWrap: "wrap",
             wordWrap: "break-word",
           }}
